move track, channel_pool, instruments to protocol/
Track now uses TYPE_CHECKING for composition/ imports —
zero runtime dependency on composition layer. Only Note
and Key imported at runtime (for isinstance checks).
