+++ midigen/__init__.py
@@ -11,9 +11,9 @@ from .composition.drums import DrumKit, Drum
 from .composition.melody import Melody
 from .composition.section import Section
 
-from .track import Track
-from .channel_pool import ChannelPool, ChannelExhaustedError
-from .instruments import INSTRUMENT_MAP
+from .protocol.track import Track
+from .protocol.channel_pool import ChannelPool, ChannelExhaustedError
+from .protocol.instruments import INSTRUMENT_MAP
 
 from .midigen import MidiGen
 from .song import Song
