+++ midigen/midigen.py
@@ -2,7 +2,7 @@ from typing import Tuple
 import os
 from mido import MidiFile
 from midigen.theory.key import Key
-from midigen.track import Track
+from midigen.protocol.track import Track
 from pathlib import Path
