+++ midigen/tests/test_track.py
@@ -3,7 +3,7 @@ from midigen.theory.note import Note
 from midigen.theory.key import KEY_MAP
 from midigen.composition.chord import Chord
 from midigen.composition.arpeggio import Arpeggio
-from midigen.track import MAX_MIDI_TICKS
+from midigen.protocol.track import MAX_MIDI_TICKS
 import unittest
